Revert change of tid from long to int
Darwin and Mach use a 32-bit int thread ID, while FreeBSD uses a long. I
changed to an int for compatibility but it caused many subtle issues. We'll
find another way to do this.

+++ contrib/llvm-project/compiler-rt/lib/safestack/safestack_platform.h
@@ -70,7 +70,7 @@ inline ThreadId GetTid() {
   DEFINE__REAL(int, _lwp_self);
   return _REAL(_lwp_self);
 #elif SANITIZER_FREEBSD
-  int Tid;
+  long Tid;
   thr_self(&Tid);
   return Tid;
 #else
